Update clean db script

+++ scripts/task_clean_database.py
@@ -7,13 +7,14 @@ from config import conda_sh_path, repo_path, db_path, clean_db_path, locations_d
 work_dir = repo_path + '/clean_database'
 
 os.environ["RAW_DB_PATH"] = db_path
-os.environ["CLEAN_DB_PATH"] = clean_db_path
+os.environ["CLEAN_DB_PATH"] =  clean_db_path
 os.environ["LOCATIONS_DB_PATH"] = locations_db_path
 os.environ["PANGO_PATH"] = pango_merged_file
 os.environ["CLADES_PATH"] = clades_merged_file
 os.environ["CLADE_CONFIG_PATH"] = clade_config_path
 os.environ["PANGO_CONFIG_PATH"] = pango_config_path
 
-os.unlink(clean_db_path)
+if os.path.exists(clean_db_path):
+    os.unlink(clean_db_path)
 out = os.system('bash -c "source ' + conda_sh_path + ' && cd ' + work_dir + ' && conda activate crs19 && python script.py"')
 sys.exit(out >> 8)
